backend/src/test/java/com/kmbbj/backend/jwt/Controller/RefreshControllerTest.java
package com.kmbbj.backend.jwt.Controller;

import com.kmbbj.backend.global.config.exception.ApiException;
import com.kmbbj.backend.global.config.exception.ExceptionEnum;
import com.kmbbj.backend.global.config.jwt.controller.RefreshController;
import com.kmbbj.backend.global.config.jwt.service.TokenService;
import com.kmbbj.backend.global.config.jwt.util.JwtTokenizer;
import io.jsonwebtoken.Claims;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.servlet.http.HttpServletResponse;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.Mockito.*;

class RefreshControllerTest {

    @Mock
    private JwtTokenizer jwtTokenizer;

    @Mock
    private TokenService tokenService;

    @InjectMocks
    private RefreshController refreshController;

    @Mock
    private HttpServletRequest request;

    @Mock
    private HttpServletResponse response;

    @Mock
    private Claims claims;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    /**
     * 리프레시 토큰이 존재하지 않을 때의 시나리오를 테스트합니다.
     */
    @Test
    void refreshTokensNoRefreshToken() {
        when(request.getHeader("Authorization")).thenReturn(null);

        ApiException exception = assertThrows(ApiException.class, () -> refreshController.refreshTokens(request, response));

        assertEquals(ExceptionEnum.TOKEN_NOT_FOUND, exception.getException());
    }
}